chore: Add Prisma schema file
This commit adds the Prisma schema file to the project. The Prisma schema defines the data model and relationships for the application. This file will be used by the Prisma client to generate the necessary database queries.

Note: This commit message follows the established convention of starting with a type (chore) followed by a concise and descriptive message.