feat: Add seeder file for generating random users and todos
This commit adds a seeder file to the project that generates random users and todos using the faker library. The seeder file creates 10 random users and 10 random todos, associating each todo with a random user. This will be used to populate the database with dummy data for testing and development purposes.

Note: This commit message follows the established convention of starting with a type (feat) followed by a concise and descriptive message.