refactor: Add user filter to PostsController
This commit adds a user filter to the PostsController in order to retrieve posts associated with a specific user. The filter is based on the user's apiKey, which is extracted from the request object. This enhancement improves the functionality of the PostsController by allowing users to view their own posts.